Separate FileOperations and DatasetOperations — no more inheritance
FileOperations(BaseResource): get(), list(), get_download_link() — /files/* only
DatasetOperations(BaseResource): all dataset-scoped methods moved from FileOperations:
  add_file_to_dataset, upload helpers, get_download_links, _fetch_files, download,
  get_thumbnails, add_thumbnail, delete_thumbnail,
  get_ingestion_requests, get_request_status, update_ingestion_status

CLI: get_file → client.files.get(), list_files(dsid) → get_associated_files()
