+++ crucible/cli/helpers.py
@@ -188,7 +188,7 @@ def cache_resource(shell_state, client, data, rtype, resource_id, **flags):
         pool = ThreadPoolExecutor(max_workers=3, thread_name_prefix='prefetch')
         futures = {
             '_keywords_future': pool.submit(client.datasets.get_keywords, resource_id),
-            '_files_future':    pool.submit(client.datasets.list_files, resource_id),
+            '_files_future':    pool.submit(client.datasets.get_associated_files, resource_id),
             '_dl_links_future': pool.submit(client.datasets.get_download_links, resource_id),
         }
         if not data.get('links'):
