+++ crucible/resources/__init__.py
@@ -8,7 +8,7 @@ while maintaining backward compatibility with the original flat API.
 """
 
 from .files import FileOperations
-from .datasets import DatasetOperations
+from .datasets import DatasetOperations  # no longer inherits FileOperations
 from .samples import SampleOperations
 from .projects import ProjectOperations
 from .users import UserOperations
